versione 0.5 Corsa,movimenti e cura sistemati,implementato HUD con vita,numero pozioni,consumo e ripristino di stamina. Adesso devo rimuovere i magic numbers sparsi per il codice

+++ Mario.h
@@ -25,7 +25,7 @@ public:
 
     void setPotionNum(int potionNum);
 
-    void recoverStam(int stamina);//ripresa stam nel tempo
+    void recoverStam(float stamina);//ripresa stam nel tempo
 bool recoverHp(int potionUsed);
 private:
     float stamina;
